code cleaned up

+++ RESTik/src/main/java/restik/view/TabloGotovnosty.java
@@ -1,7 +1,7 @@
 package restik.view;
 
 public class TabloGotovnosty {
-    public static void displayReadyTime(String orderId, Integer time) {
+    public static void displayReadyTime(String orderId, Double time) {
         System.out.println("Order " + orderId + " will be ready in " + time + " minutes");
     }
 }
